perf: convert signal cache to float once at load, pass float* to ImPlot
ImPlot renders using float vertices internally. Previously all signal
data was stored as double and ImPlot was converting double→float on
every PlotLine call (N samples × K components × 60fps).

Changes:
- PlottedSignal gains time_cache_f (vector<float>) and changes
  component_cache to vector<vector<float>>
- build_component_cache fills both time and value caches as float,
  converting from the double SignalBuffer exactly once at load time
- Scalar signals are now also stored in component_cache[0], making
  the render loop uniform (no special-case for n_components==1)
- trajectory_2d_plot and ground_track_plot updated to float caches
- Per-frame double→float conversion eliminated entirely

+++ include/gnc_viz/ground_track_plot.hpp
@@ -20,8 +20,8 @@ public:
 private:
     // Pre-computed lat/lon per signal (plot_key → cache). Valid while buffer ptr unchanged.
     struct GroundTrackCache {
-        std::vector<double>          lats;
-        std::vector<double>          lons;
+        std::vector<float>           lats;
+        std::vector<float>           lons;
         std::weak_ptr<SignalBuffer>  source;
     };
     std::unordered_map<std::string, GroundTrackCache> m_cache;
